globally enable maven stacktrace processing.
