+++ java/maven/src/org/netbeans/modules/maven/output/ExecPluginOutputListenerProvider.java
@@ -19,64 +19,25 @@
 package org.netbeans.modules.maven.output;
 
 import org.netbeans.modules.maven.NbMavenProjectImpl;
-import org.netbeans.modules.maven.api.output.OutputProcessor;
-import org.netbeans.modules.maven.api.output.OutputUtils;
-import org.netbeans.modules.maven.api.output.OutputVisitor;
-import org.netbeans.api.java.classpath.ClassPath;
-import org.netbeans.api.project.Project;
-import org.netbeans.modules.maven.api.classpath.ProjectSourcesClassPathProvider;
-import org.netbeans.spi.java.classpath.support.ClassPathSupport;
-import org.openide.windows.OutputListener;
-
-
 
 
 /**
  * exec plugin output processing, just handle stacktraces.
  * @author  Milos Kleint
+ * 
+ * @deprecated use JavaStacktraceOutputListenerProvider instead
  */
-public class ExecPluginOutputListenerProvider implements OutputProcessor {
-    
+public class ExecPluginOutputListenerProvider extends JavaStacktraceOutputProcessor {
+
     private static final String[] EXECGOALS = new String[] {
         "mojo-execute#exec:exec", //NOI18N
         "mojo-execute#exec:java", //NOI18N
         "mojo-execute#javafx:run" //NOI18N
     };
-    private final NbMavenProjectImpl project;
-    
+
     /** Creates a new instance of ExecPluginOutputListenerProvider */
     public ExecPluginOutputListenerProvider(NbMavenProjectImpl proj) {
-        project = proj;
-    }
-    
-    @Override
-    public void processLine(String line, OutputVisitor visitor) {
-        OutputVisitor.Context context = visitor.getContext();
-        Project prj = project;
-        if (context != null && context.getCurrentProject() != null) {
-            prj = context.getCurrentProject();
-        }
-        OutputListener list = OutputUtils.matchStackTraceLine(line, prj);
-        if (list != null) {
-            visitor.setOutputListener(list);
-        }
+        super(proj, EXECGOALS);
     }
 
-    @Override
-    public String[] getRegisteredOutputSequences() {
-        return EXECGOALS;
-    }
-
-    @Override
-    public void sequenceStart(String sequenceId, OutputVisitor visitor) {
-    }
-
-    @Override
-    public void sequenceEnd(String sequenceId, OutputVisitor visitor) {
-    }
-    
-    @Override
-    public void sequenceFail(String sequenceId, OutputVisitor visitor) {
-    }
-    
 }
